SIL: Stop using TypeConverter::getArchetypes() in type lowering, NFC
Now that we can ask questions of a generic signature directly,
we can pass dependent types right on through to TypeClassifierBase
without consulting getArchetypes().

This required a small fix to getTypeLowering() to correctly form
an abstraction pattern from an interface type.

There are no remaining usages of getArchetypes(), so that
entire ArchetypeBuilder instance is now gone.

Also refactor away SIL's take on the GenericsRAII helper class.