fix <rdar://problem/24221830> Bogus "no calls to throwing functions" warning in derived throwing init
This is a case where we used to produce:

<unknown>:0: warning: no calls to throwing functions occur within 'try' expression

Which is bogus, due to the try expr implicitly generated as part of the
implicit super.init call for an init that doesn't otherwise contain a super.init.
Silence this warning by ignoring implicitly generated trys, since this try gets
produced before name binding has resolved exactly which try is being invoked.