[Sema] Fix assertion building refinement context for synthesized accessors.
For VarDecls with synthesized trivial accessors, we may have not a valid
location for the end of the braces, which causes an assertion failure when
constructing the type refinement context range for underlying storage.

This is a quick fix to fall back to using the range of the storage in this case.
The right fix here is to update AbstractStorageDecl::addTrivialAccessors() to
take brace locations and have callers of that method provide appropriate source
locations.

rdar://problem/24258839