Don't add capture arguments to default argument generators.
If we ever add the ability for default arguments to refer to
enclosing variables, we should promote DefaultArgumentInitializer
to be an always-present entity, give it an independent capture
list, and make SILDeclRef refer to it directly.  In fact, we may
want to do that anyway.  Until then, it's a weird special case
in terms of not really being a reference to the function returned
by getAnyFunctionRef().

rdar://24242783