Sema: Refactor away SelfInResultType, NFC
SelfInResultType::DynamicSelf was unnecessary because we already
returned SelfUse::Result and handled it correctly if the witness
had a dynamic Self return type.

SelfInResultType::Ignore was only used for constructors, but it
is better to handle them just like FuncDecls with dynamic Self
returns.

Also, the additional VarDecl case was unnecessary.