Fix bitcasting issue in partial apply forwarding thunks.
In the case where we generate a forwarding thunk with no polymorphic
arguments that forwards to a method that has polymorphic arguments, we
were grabbing the wrong type for the cast of the context argument. The
context argument, if there is one, should be after the polymorphic
arguments.

Fixes rdar://problem/23528633.

(cherry picked from commit 006637f98a46ebf851985b9a7c32110497292926)