test: remove `ulimit -c unlimited`
The core limit adjustment was added in SVN r17964 (fad87470) by Ted Kremenek.
This is problematic in bash where if a `ulimit -c 0` was invoked prior to the
test running, it would fail due to the failure to set the core limit to
unlimited.  Interestingly enough, this will be different across bash and zsh.